requirements.txt
# Web scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# Environment and configuration
python-dotenv==1.0.1
PyYAML==6.0.1

# Data handling
pandas==2.2.0

# Utilities
tqdm==4.66.1  # Progress bars
retrying==1.3.4  # Retry logic
python-dateutil==2.8.2

# Logging
colorlog==6.8.0

# Testing (optional)
pytest==8.0.0
pytest-mock==3.12.0

# Linting (optional)
flake8==7.0.0
black==24.1.1